requirements.txt
python-dateutil~=2.8.1
influxdb~=5.3.1
Flask~=1.1.2